+++ src/hooks/useRequest.js
@@ -18,21 +18,22 @@ const useRequest = (
   url,
   { isCacheResponse, params, responseDataConverter, ...config } = {}
 ) => {
-  const context = useSadnessContext();
-  const [response, setResponse] = useState(EMPTY_RESPONSE);
-
-  const extra = { isCacheResponses: isCacheResponse };
   const request = toRequestRecord({ ...config, params, url });
+  const emptyResponse = EMPTY_RESPONSE.set("request", request);
+  const extra = { isCacheResponses: isCacheResponse };
 
-  const handleAPIRequest = (
+  const context = useSadnessContext();
+  const [response, setResponse] = useState(emptyResponse);
+
+  const handleAxiosRequest = ({
     isCheckCache = false,
     isSetEmptyResponseOnStart = false
-  ) => {
+  } = {}) => {
     const { axios, onErrorRequest, onStartRequest, onSuccessRequest } = context;
     onStartRequest();
 
     if (isSetEmptyResponseOnStart) {
-      setResponse(EMPTY_RESPONSE);
+      setResponse(emptyResponse);
     }
 
     if (isCheckCache) {
@@ -64,18 +65,26 @@ const useRequest = (
   };
 
   useEffect(() => {
-    handleAPIRequest(true);
+    handleAxiosRequest({ isCheckCache: true });
   }, []);
 
   return new ResponseContext({
     onReload: (isFullReload = false) => {
-      handleAPIRequest({
+      handleAxiosRequest({
         isCheckCache: false,
         isSetEmptyResponseOnStart: isFullReload
       });
     },
-    onUpdate: nextState => {
-      setResponse(nextState);
+    onUpdate: (nextResponse = null) => {
+      if (!nextResponse) {
+        setResponse(emptyResponse);
+      } else if (nextResponse.status >= 400) {
+        setResponse(toErrorResponseRecord(request, { response: nextResponse }));
+      } else {
+        setResponse(
+          toSuccessResponseRecord(request, nextResponse, responseDataConverter)
+        );
+      }
     },
     response
   });