+++ stories/settings.js
@@ -1,0 +1,2 @@
+export const API_URL = process.env.API_URL || "https://igordavydenko.com/api/";
+export const UI_URL = process.env.UI_URL || "https://igordavydenko.com";